requirements.txt
tweepy
python-dotenv
pyppeteer